federation/coordination-hub/grpc_server.py
"""
gRPC server for federation gossip and node-to-node communication
"""
import asyncio
import logging
from concurrent import futures
import grpc
from datetime import datetime

# Placeholder for generated protobuf code
# In production, generate from .proto files
logger = logging.getLogger(__name__)


class FederationServicer:
    """gRPC servicer for federation protocol"""

    def __init__(self, hub):
        self.hub = hub

    async def Gossip(self, request, context):
        """Handle gossip message"""
        logger.info(f"Received gossip from {request.sender_node_id}: {request.message_type}")

        # Process message based on type
        if request.message_type == "heartbeat":
            # Update heartbeat
            pass

        elif request.message_type == "vote_initiated":
            # Notify local node of new vote
            pass

        elif request.message_type == "node_join":
            # New node joined federation
            pass

        return {"status": "ok"}

    async def RequestVote(self, request, context):
        """Raft: Handle vote request"""
        logger.info(f"Vote request from {request.candidate_id} for term {request.term}")

        # Forward to consensus engine
        # Return vote response

        return {
            "term": request.term,
            "vote_granted": True
        }

    async def AppendEntries(self, request, context):
        """Raft: Handle append entries (heartbeat + log replication)"""
        logger.debug(f"AppendEntries from {request.leader_id} (term {request.term})")

        # Forward to consensus engine

        return {
            "term": request.term,
            "success": True
        }


async def serve_grpc(hub, port: int, tls_cert: str = None, tls_key: str = None):
    """Start gRPC server"""
    server = grpc.aio.server(futures.ThreadPoolExecutor(max_workers=10))

    # Add servicer
    servicer = FederationServicer(hub)
    # federation_pb2_grpc.add_FederationServiceServicer_to_server(servicer, server)

    # Configure TLS if enabled
    if tls_cert and tls_key:
        with open(tls_cert, 'rb') as f:
            cert = f.read()
        with open(tls_key, 'rb') as f:
            key = f.read()

        credentials = grpc.ssl_server_credentials([(key, cert)])
        server.add_secure_port(f'[::]:{port}', credentials)
        logger.info(f"gRPC server listening on port {port} (TLS enabled)")
    else:
        server.add_insecure_port(f'[::]:{port}')
        logger.info(f"gRPC server listening on port {port} (insecure)")

    await server.start()
    await server.wait_for_termination()
