federation/coordination-hub/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
python-multipart==0.0.6
httpx==0.26.0
prometheus-client==0.19.0
